Moved classes to their own file and added missing XML comments
