+++ src/MefContrib.Hosting.Conventions/MefContrib.Hosting.Conventions.csproj
@@ -83,10 +83,15 @@
     <Compile Include="Configuration\ImportConventionBuilder.cs" />
     <Compile Include="Configuration\ImportRegistry.cs" />
     <Compile Include="Configuration\ITypeDefaultConvention.cs" />
+    <Compile Include="Configuration\ITypeDefaultConventionBuilder.cs" />
+    <Compile Include="Configuration\ITypeDefaultConventionConfigurator.cs" />
     <Compile Include="Configuration\PartConventionBuilderExtensions.cs" />
     <Compile Include="Configuration\IPartRegistry.cs" />
     <Compile Include="Configuration\PartConventionBuilder.cs" />
     <Compile Include="Configuration\PartRegistry.cs" />
+    <Compile Include="Configuration\TypeDefaultConvention.cs" />
+    <Compile Include="Configuration\TypeDefaultConventionBuilder.cs" />
+    <Compile Include="Configuration\TypeDefaultConventionConfigurator.cs" />
     <Compile Include="ContractService.cs" />
     <Compile Include="ConventionContractService.cs" />
     <Compile Include="ConventionCatalog.cs" />
